refactor(grpc-sdk)!: remove module management specifics and keep sdk-only behaviour feat(module-tools): new module to handle module-specific operations
