+++ libraries/grpc-sdk/src/classes/ConduitAuthorizedResource.ts
@@ -1,4 +1,4 @@
-import { Resource_Permission, Resource_Relation } from '../protoUtils/authorization';
+import { Resource_Relation, Resource_Permission } from '../protoUtils/authorization';
 
 export class ConduitAuthorizedResource {
   readonly name: string;
