+++ libraries/grpc-sdk/src/classes/index.ts
@@ -1,9 +1,5 @@
 export * from './Error';
+export * from './ConduitModule';
+export * from './HealthCheck';
 export * from './ConduitSchema';
-export * from './ConduitActiveSchema';
-export * from './GrpcServer';
-export * from './ConduitServiceModule';
-export * from './ManagedModule';
-export * from './ModuleManager';
-export * from './ConfigController';
 export * from './ConduitAuthorizedResource';
