+++ libraries/grpc-sdk/src/interfaces/ConduitRoute.ts
@@ -1,9 +1,8 @@
-import { ConduitRouteReturnDefinition } from '../ConduitRouteReturn';
+import { ConduitRouteReturnDefinition } from './ConduitRouteReturn';
 import { ConduitRouteOptions } from './Route';
-import { RequestHandlers } from '../wrapRouterFunctions';
 
 export interface ConduitRoute {
   returnType: ConduitRouteReturnDefinition;
   input: ConduitRouteOptions;
-  handler: RequestHandlers;
+  handler: any;
 }
