+++ libraries/grpc-sdk/src/interfaces/ConduitSchemaExtension.ts
@@ -1,4 +1,4 @@
-import { ConduitModel } from '.';
+import { ConduitModel } from './Model';
 
 export interface ConduitSchemaExtension {
   readonly schemaName: string;
