+++ libraries/grpc-sdk/src/interfaces/index.ts
@@ -1,8 +1,13 @@
-export * from './ConduitSchemaExtension';
-export * from './ConduitService';
+export * from './ConduitMiddleware';
+export * from './ConduitRoute';
+export * from './ProxyRoute';
+export * from './Route';
+export * from './Socket';
+export * from './types';
 export * from './Model';
 export * from './GrpcCallback';
 export * from './Indexable';
 export * from './RawQuery';
 export * from './UntypedArray';
 export * from './Params';
+export * from './ConduitRouteReturn';
