+++ libraries/grpc-sdk/src/metrics/MetricsServer.ts
@@ -1,7 +1,7 @@
-import ConduitGrpcSdk from '../index';
 import express from 'express';
 import { Registry } from 'prom-client';
 import { isNaN } from 'lodash';
+import ConduitGrpcSdk from '../index';
 
 export class MetricsServer {
   private readonly moduleName: string;
