+++ libraries/grpc-sdk/src/metrics/clientMiddleware.ts
@@ -1,5 +1,4 @@
 // Forked from nice-grpc-prometheus
-
 import { isAbortError } from 'abort-controller-x';
 import {
   CallOptions,
