+++ libraries/grpc-sdk/src/modules/database/index.ts
@@ -1,12 +1,13 @@
-import { ConduitModule } from '../../classes/ConduitModule';
-import { ConduitSchema } from '../../classes';
+import { ConduitModule } from '../../classes';
 import {
   DatabaseProviderDefinition,
   DropCollectionResponse,
   Schema,
 } from '../../protoUtils/database';
-import { ConduitSchemaExtension, RawQuery, UntypedArray } from '../../interfaces';
+import { RawQuery, UntypedArray } from '../../interfaces';
 import { Query } from '../../types/db';
+import { ConduitSchema } from '../../classes/ConduitSchema';
+import { ConduitSchemaExtension } from '../../interfaces/ConduitSchemaExtension';
 
 export class DatabaseProvider extends ConduitModule<typeof DatabaseProviderDefinition> {
   constructor(private readonly moduleName: string, url: string, grpcToken?: string) {
