--- libraries/grpc-sdk/src/routing/interfaces/index.ts
@@ -1,6 +1,0 @@
-export * from './ConduitMiddleware';
-export * from './ConduitRoute';
-export * from './Route';
-export * from './Socket';
-export * from './types';
-export * from './ProxyRoute';
