+++ libraries/module-tools/README.mdx
@@ -1,0 +1,21 @@
+# Conduit gRPC SDK
+
+This library provides a framework to bind all Conduit modules in a safe and consistent way.<br />
+It is also used to provide syntactic sugar that makes development easier.
+
+## Build Steps
+
+- run build.sh
+- run npm build
+
+## Features ✔️
+
+- Service registration & discovery
+- Application route registration
+- Administrative route registration
+- Provides wrappers for gRPC handlers to support async
+- Parses incoming requests and outgoing responses to JSON
+- Provides health check service out of the box
+- Manages the gRPC server so you don't have to
+- Provides typings for common Conduit models
+- Provides classes to facilitate building your own Conduit modules
