+++ libraries/module-tools/package.json
@@ -1,0 +1,65 @@
+{
+  "name": "@conduitplatform/module-tools",
+  "version": "1.1.3",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "engines": {
+    "node": ">=14"
+  },
+  "scripts": {
+    "prepublish": "npm run build",
+    "prebuild": "npm run protoc",
+    "build": "rimraf dist && tsc",
+    "postbuild": "copyfiles -u 1 src/*.proto src/**/*.json ./dist/",
+    "protoc": "sh build.sh"
+  },
+  "license": "MIT",
+  "dependencies": {
+    "@conduitplatform/grpc-sdk": "^1.0.0",
+    "@grpc/grpc-js": "^1.6.7",
+    "@grpc/proto-loader": "^0.6.13",
+    "@types/google-protobuf": "^3.15.6",
+    "@types/ioredis": "^4.28.10",
+    "@types/lodash": "^4.14.182",
+    "abort-controller-x": "^0.4.0",
+    "convict": "^6.2.3",
+    "express": "^4.18.1",
+    "fast-jwt": "^1.6.0",
+    "fs-extra": "^10.1.0",
+    "ioredis": "^5.1.0",
+    "lodash": "^4.17.21",
+    "prom-client": "^14.0.1",
+    "protobufjs": "^6.11.3",
+    "winston": "^3.8.1",
+    "winston-loki": "^6.0.5"
+  },
+  "directories": {
+    "lib": "dist",
+    "src": "src"
+  },
+  "files": [
+    "dist"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "ssh://git@github.com/ConduitPlatform/conduit.git",
+    "directory": "libraries/module-tools"
+  },
+  "devDependencies": {
+    "@types/convict": "^6.1.1",
+    "@types/express": "^4.17.13",
+    "@types/fs-extra": "^9.0.13",
+    "@types/google-protobuf": "^3.15.6",
+    "@types/ioredis": "^4.28.10",
+    "@types/lodash": "^4.14.182",
+    "@types/node": "14.14.31",
+    "copyfiles": "^2.4.1",
+    "rimraf": "^3.0.2",
+    "ts-proto": "^1.117.0",
+    "typescript": "~4.7.4"
+  },
+  "peerDependencies": {
+    "@grpc/grpc-js": "^1.5.2",
+    "@grpc/proto-loader": "^0.6.13"
+  }
+}
