+++ libraries/module-tools/src/classes/ConfigController.ts
