+++ libraries/module-tools/src/classes/index.ts
@@ -1,0 +1,6 @@
+export * from './ConduitActiveSchema';
+export * from './GrpcServer';
+export * from './ConduitServiceModule';
+export * from './ManagedModule';
+export * from './ModuleManager';
+export * from './ConfigController';
