+++ libraries/module-tools/src/index.ts
@@ -1,0 +1,4 @@
+export * from './interfaces';
+export * from './classes';
+export * from './helpers';
+export * from './routing';
