+++ libraries/module-tools/src/routing/index.ts
@@ -1,5 +1,3 @@
-export * from './interfaces';
-export * from './ConduitRouteReturn';
 export * from './RouteBuilder';
 export * from './RoutingManager';
 export * from './RoutingUtilities';
